implemented search for firstName, lastName, and email
